# api_swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a Scrum project management server, with the multiple services provided by this server.",
    "version" : "1.0.0",
    "title" : "Bepp Project",
    "contact" : {
      "email" : "mathieuperez3@gmail.com"
    }
  },
...
  "schemes" : [ "http" ],
  "paths" : {
    "/api/users/" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create user",
        "description" : "This can only be done by the logged in user.",
        "operationId" : "createUser",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Created user object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid username/password supplied"
          }
        }
      }
    },
    "/api/users/token" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Logs user into the system",
        "description" : "",
        "operationId" : "loginUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "login",
          "in" : "query",
          "description" : "The user name for login",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "The password for login in clear text",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid username/password supplied"
          }
        }
      }
    },
    "/api/users/{login}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Get user by login",
        "description" : "",
        "operationId" : "getUserByName",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "login",
          "in" : "path",
          "description" : "The name that needs to be fetched.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    },
...
    "/api/projects/{name}/users/{login}" : {
      "put" : {
        "tags" : [ "project" ],
        "summary" : "Add user to project",
        "description" : "",
        "operationId" : "addUser",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "The user name for login",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "login",
          "in" : "path",
          "description" : "The password for login in clear text",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "role",
          "in" : "query",
          "description" : "The role for the user in the project",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "403" : {
            "description" : "User not allowed"
          },
          "404" : {
            "description" : "Project or User not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "petstore_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "http://petstore.swagger.io/oauth/dialog",
      "flow" : "implicit",
      "scopes" : {
        "write:pets" : "modify pets in your account",
        "read:pets" : "read your pets"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
    "User" : {
      "type" : "object",
      "properties" : {
        "login" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "User"
      }
    },
...
  }
}